arch: fix config module — extract paths, move MCP description (#144)
- Extract path helpers (state_dir, log_dir, reminders_dir, agent_bus_socket)
  to new paths.rs module; re-export from config for backward compat
- Move send_message_description() from config::UserConfig to mcp module
  as build_send_message_description() free function
- Fix TaskCriteria import: config now references crate::domain::task
  directly instead of crate::task (explicit layer dependency)
- config.rs reduced from 866 to 754 lines
